refactor: 훅 다시 걷어내기..
