+++ src/components/Product/ProductList/ProductList.tsx
@@ -1,10 +1,19 @@
 import * as styles from './ProductList.style';
-import useProductList from '../../../hooks/useProductList';
 import Dropdown from '../../Dropdown/Dropdown';
 import Spinner from '../../Spinner/Spinner';
 import ProductCard from '../ProductCard/ProductCard';
 import { CATEGORY_OPTIONS, ORDER_BY_OPTIONS } from '../../../constants/categoryOption';
 import { CartItem } from '../../../types/cartContents';
+import { useCallback, useState } from 'react';
+import { CategoryOptionType, OrderByOptionType } from '../../../types/categoryOption';
+import { useErrorContext } from '../../../contexts/ErrorContext';
+import useFetch from '../../../hooks/useFetch';
+import { ProductResponse } from '../../../types/response';
+import getProducts from '../../../api/getProducts';
+import useErrorHandler from '../../../hooks/useErrorHandler';
+import { createProductListViewModel, ProductCardViewModel } from '../../../api/model/createProductListModel';
+import { deleteCartItem } from '../../../api/deleteCartItem';
+import postCartItem from '../../../api/postCartItem';
 
 interface ProductListProps {
   cartItems?: CartItem[];
@@ -13,8 +22,45 @@ interface ProductListProps {
 }
 
 export default function ProductList({ cartItems, refetchCart, cartItemsLoading }: ProductListProps) {
-  const { category, setCategory, orderBy, setOrderBy, productListViewModel, productFetchLoading, handleCartToggle } =
-    useProductList({ cartItems, refetchCart });
+  const [category, setCategory] = useState<CategoryOptionType>('전체');
+  const [orderBy, setOrderBy] = useState<OrderByOptionType>('낮은 가격순');
+  const { showError } = useErrorContext();
+
+  const {
+    data: products,
+    isLoading: productFetchLoading,
+    error: productFetchError
+  } = useFetch<ProductResponse>({
+    fetchFn: getProducts(orderBy),
+    deps: [orderBy]
+  });
+
+  useErrorHandler(productFetchError);
+
+  const filteredProducts =
+    category === '전체' ? products?.content : products?.content.filter((item) => item.category === category);
+
+  const productListViewModel = createProductListViewModel({
+    products: filteredProducts,
+    cartItems
+  });
+
+  const handleCartToggle = useCallback(
+    async (product: ProductCardViewModel) => {
+      try {
+        if (product.isInCart) {
+          await deleteCartItem(product.cartItemId!);
+          await refetchCart();
+        } else {
+          await postCartItem(product.id, 1);
+          await refetchCart();
+        }
+      } catch (err) {
+        if (err instanceof Error) showError(err);
+      }
+    },
+    [refetchCart, showError]
+  );
 
   return (
     <>
